[#104] move mixin
